Add integration specs
This commit also adds a way to run the integration specs asynchronously
for the sake of performance as a proof of concept. There are literally
only two specs touching the DB at this point so it's not intended to be
a serious solution to anything.

It monkeypatches Spec so it's probably gonna get weird at some point. It
is intended to be a potential upstream patch to Crystal, so if this ends
up not being terrible, we'll open a PR for it then. If it gets weird
before that, we'll probably just remove it. Either way, it's only
temporary code in this repo.