Handle signed ints from -0x19...0x10
This was from some copypasta from the msgpack-crystal shard. I'm not
sure if the MessagePack optimization there even works with Neo4j, but
this commit just removes the attempt at it for now. Saving 1 byte here
and there won't be the most amazing performance boost ever, so I'm okay
with removing the optimization until I understand it better.