Add automatic casting of int sizes
Neo4j sends us data in the smallest size the value fits in. For example,
if the value is 42, it will serialize it as an 8-bit signed int. The
problem was that if we tried to map it to a different type of int, it
raised an exception. This commit handles the type casting for you.