Improve connection self-healing
If the DB connection died, we recursively reinitialize, but this had two
problems:

- We weren't handling all of the exceptions that could be raised by a
  dead connection
- We didn't limit how many times we would reinitialize the connection on
  a single query attempt

Handling all of the exceptions is an obvious requirement for
self-healing, but the retry limit is there in case it can't self-heal at
all and needs to start yelling loudly. You can't hug every cat and you
can't heal every connection.