Move test models to top-level namespace
The fact that they were nested under the Neo4j namespace made it so the
macros worked when not fully qualified, but in most apps the models
won't be namespaced under Neo4j (or at least, they shouldn't be) so this
commit fixes that. Not having fully qualified class references inside
the macro code was causing problems.