Allow passing a block to execute and exec_cast
This gives the ability to stream results in through these methods rather
than having to contain the full list of results in memory.

This differs from the stream method in that it doesn't return an
Iterable. Turns out that that is kinda dangerous because the following
scenario is too easy to fall into:

    def run_query(query)
      pool.connection do |connection|
        connection.stream(query)
      end
    end

In this example, we return the unconsumed response and then return the
connection to the connection pool for another fiber to pick up. There is
no guarantee that the response will be consumed before that connection
is checked out from the pool again.

The block variants of execute and exec_cast consume the entire response
before they return, so they protect against that scenario while still
reducing the latency to the first result, which was the impetus behind
the stream method.

This commit also contains the following changes:
- Extracted all Connection#write_message calls to a Connection#send
  method. All calls inside Connection use this method now.
- Neo4j::ValueType and Neo4j::ResponseType no longer include Type in
  their names. They're just Neo4j::Value and Neo4j::Response.
- Fixed the user agent declaration to use the actual version of the
  shard.